renamed P2PService to BeaconNodeService
