+++ services/beaconchain/build.gradle
@@ -12,7 +12,7 @@ dependencies {
   implementation project(':services')
   implementation project(':services:powchain')
   implementation project(':pow')
-  implementation project(':services:p2p')
+  implementation project(':services:beaconnode')
 
   implementation 'com.google.guava:guava'
   implementation 'org.apache.logging.log4j:log4j-api'
