+++ settings.gradle
@@ -5,7 +5,7 @@ include 'networking:p2p'
 include 'services'
 include 'services:beaconchain'
 include 'services:powchain'
-include 'services:p2p'
+include 'services:beaconnode'
 include 'services:chainstorage'
 include 'artemis'
 include 'util'
